event_consumer/requirements.txt
aio-pika==9.1.1
python-dotenv==1.0.0
aioredis==2.0.1
